Refactor ObsDim out of LearnBase. Also updates to getobs interface.
